+++ e2e/e2etests/test_zeta_deposit_and_call.go
@@ -36,7 +36,7 @@ func TestZetaDepositAndCall(r *runner.E2ERunner, args []string) {
 
 	// wait for the cctx to be mined
 	cctx := utils.WaitCctxMinedByInboundHash(r.Ctx, tx.Hash().Hex(), r.CctxClient, r.Logger, r.CctxTimeout)
-	r.Logger.CCTX(*cctx, "deposit_and_call")
+	r.Logger.CCTX(*cctx, "zeta_deposit_and_call")
 	require.Equal(r, crosschaintypes.CctxStatus_OutboundMined, cctx.CctxStatus.Status)
 
 	// check the payload was received on the contract