+++ e2e/e2etests/test_zeta_deposit_and_call_revert.go
@@ -34,7 +34,7 @@ func TestZetaDepositAndCallRevert(r *runner.E2ERunner, args []string) {
 
 	// wait for the cctx to be reverted
 	cctx := utils.WaitCctxMinedByInboundHash(r.Ctx, tx.Hash().Hex(), r.CctxClient, r.Logger, r.CctxTimeout)
-	r.Logger.CCTX(*cctx, "deposit_and_call")
+	r.Logger.CCTX(*cctx, "zeta_deposit_and_call")
 	require.Equal(r, crosschaintypes.CctxStatus_Reverted, cctx.CctxStatus.Status)
 
 	// check the balance is more than 0