+++ changelog.md
@@ -7,6 +7,7 @@
 * [3929](https://github.com/zeta-chain/node/pull/3929) - add support for TON http-rpc
 * [3958](https://github.com/zeta-chain/node/pull/3958) - integrate TON http-rpc
 * [3961](https://github.com/zeta-chain/node/pull/3961) - cosmos v53 upgrade
+* [3990](https://github.com/zeta-chain/node/pull/3990) - add support for deposits using the new v2 connector contract
 
 ### Fixes